backend/requirements.txt
fastapi>=0.128.0
sqlmodel>=0.0.31
python-dotenv>=1.2.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.5.0
argon2-cffi>=25.1.0
email-validator>=2.3.0
python-multipart>=0.0.21
psycopg[binary]>=3.3.2
aiosqlite>=0.22.1
httpx>=0.28.1
uvicorn>=0.40.0
anthropic